fix: address ownerDocument permission error in Firefox
